Fix patching failure caused by MockMailHost import.
The imports of Products.CMFPlone.tests.utils (from which MockMailHost is imported)
implicitly import the CMFPlone patches.
Since this import happens on test discovery this can cause that certain content types
cannot be created (the factory is missing).
Therefore those imports should be inside the function and not globally, so that they
are triggered by the testing layer at the right moment.