Remove dependency to Products.PloneHotfix20121106 and tests.
The TestToolMocking test suite depends on having Products.PloneHotfix20121106,
which is bad because those hotfixes are incorporated in the newest Plone hotfix releases.
This causes the test to break on Plone 4.2 because the hotfix does no longer apply.

This removes the regression test which tests that getToolByName is mocked and released
properly even when it is patched before by a hotfix.
We cannot test this properly for all supported Plone versions because the hotfix does
not apply.
It is also not a good idea to have a dependency on a hotfix when test the latest
hotfix versions, since the required hotfixes change.