Extend "staticuid" to also be a context manager.
It should be possible to use staticuid multiple times within the same
function with different prefixes.

staticuid did only support beeing used as function decorator,
therefore this use case was only possible by using nested functions,
which is not very readable.

With this change staticuid can also be used a context manager,
while it can still be used as decorator.