Fix mailing tests:
The email messages captured by the MockMailHost are represented as bytestrings.

This is not wrong, since sending mails via Zope's MailHost is very close to
SMTP, and over SMTP one transmits *bytes*, not necessarily *text*. SMTP will
only ever send 7-bit ASCII over the network, any encoding of non-ASCII
characters happens inside the RFC 822 message, using binascii encodings like
Base64 or Quoted-Printable.

However, in Python 3 this means those messages also need to be compared to
bytestrings, and a comparison like

```
self.assertIn('foo@example', messageText)
```

will fail with

```
TypeError: a bytes-like object is required, not 'str'
```

The quick fix therefore is to rewrite the tokens we search for in the messages
(in ftw.testings tests) to bytestrings, and call it a day.

In the future we could consider whether we want to change the MockMailHost so
that it returns unicode strings, but that didn't immediately work (issues with
line endings), and also obscures what's going on.