Fix genericsetup uninstall profile test:
For Plone >= 5.2 a registry record 'custom_css_timestamp' for
plone.app.theming.interfaces.IThemeSettings exists that screwed up the
GS profile snapshot diffing.

That record has both a datetime value and a defaultFactory with datetime.now()
that need to be neutralized in order for the diff not to change all the time.